release-token: consolidate to one canonical format + gate signing on Actuate
Address PR #748 Copilot review (both comments).

#1 — duplicate/divergent release-token format. The crate previously used a
weaker payload (plain SHA-256 over the raw image, bare-digest signature, bool
verify) that diverged from the existing, disciplined `src/governor_release.rs`
(domain-separated + length-prefixed digest and signing payloads, digest-carrying
token, typed ReleaseDenied). Two token formats for the same HVCHAN §3.5-7 seam is
an integration/replay hazard. Consolidate to ONE format:
  - kirra-release-token now IS the canonical implementation (ported from
    governor_release.rs): DIGEST_DOMAIN / RELEASE_DOMAIN tags, length-prefixed
    payloads, ReleaseToken { digest, signature } (96B wire), verify_release ->
    Result<(), ReleaseDenied>, verify_strict.
  - src/governor_release.rs is now a thin re-export of the lean crate, keeping
    only the root-local governor_key_id (needs audit_chain). Root callers
    (`kirra_verifier::governor_release::*`, tests/clause2_trust_chain) unchanged.
  - root crate gains a kirra-release-token path dependency.

#2 — signing footgun. GovernorCycle.view is Some whenever a command was RECEIVED
(transport+codec ok), so it is Some even on a kinematic DenyBreach -> SafeStop.
Add GovernorCycle::view_to_sign() returning Some only when outcome is Actuate, so
"sign only when actuatable" is enforced at the seam, not by caller discipline. New
integration test proves a denied (InvalidTimeDelta) command has view.is_some() but
view_to_sign() == None.

Verified: kirra-release-token 8 unit + 3 integration, kirra-core 179+2, root
governor_release unit + clause2_trust_chain; clippy clean (lean + core).
