+++ Cargo.toml
@@ -71,6 +71,10 @@ kirra-fleet-types = { path = "crates/kirra-fleet-types" }
 # binding (src/governor_release.rs) reuses THIS crate's canonical image plus the
 # verifier's EXISTING Ed25519/SHA-256 machinery (no new crypto).
 kirra-contract-channel = { path = "crates/kirra-contract-channel" }
+# The single canonical governor→actuator release-token format (HVCHAN §3.5-7).
+# `src/governor_release.rs` re-exports it (adding only the root-local key-id
+# helper) so there is one digest/sign/verify implementation, not two.
+kirra-release-token = { path = "crates/kirra-release-token" }
 parko-core = { path = "parko/crates/parko-core" }
 # Governor-free wire schema for capture records, shared verbatim with the offline
 # collector (docs/COLLECTOR_DESIGN.md [C1]). Re-exported by `crate::capture`.
