+++ crates/kirra-release-token/src/lib.rs
@@ -1,4 +1,4 @@
-//! # kirra-release-token — the governor→actuator release bridge (HVCHAN §3.5-6)
+//! # kirra-release-token — the governor→actuator release bridge (HVCHAN §3 steps 5-7)
 //!
 //! The last link of the L3 trust chain (ADR-0030 Clause F, ADR-0013): once the
 //! governor has validated a [`GovernorContractView`] and decided it is
@@ -7,18 +7,31 @@
 //! these bytes."* A guest cannot forge it; a single flipped byte between governor
 //! and actuator invalidates it.
 //!
-//! Two steps, matching HVCHAN §3:
-//! - **step 5 — digest.** SHA-256 over the canonical little-endian image
-//!   ([`GovernorContractView::canonical_image`]).
-//! - **step 6 — sign.** Ed25519 over that digest, with the governor's key.
+//! Steps 1-4 (the frozen layout, the seqlock coherent read, and validation) live
+//! in [`kirra_contract_channel`]. This crate closes the chain:
 //!
-//! **No new crypto primitives** (HVCHAN §3): the SAME `ed25519-dalek` +
-//! `sha2` the verifier's attestation path already uses (`src/attestation.rs`),
-//! applied at this seam. `#![forbid(unsafe_code)]`.
+//! - **Step 5 — digest.** [`contract_digest`] hashes the **exact validated
+//!   snapshot bytes** ([`GovernorContractView::canonical_image`]) — the bytes the
+//!   judge approved, not the live region which may already have moved on.
+//! - **Step 6 — release token.** [`issue_release_token`] signs that digest with
+//!   the governor's Ed25519 key.
+//! - **Step 7 — actuator verify-before-release.** [`verify_release`] re-derives
+//!   the digest over the command the actuator is **about to actuate** and checks
+//!   (a) the token's digest matches it **and** (b) the signature verifies against
+//!   the governor key. Either failure ⇒ **no release** (fail-closed).
 //!
-//! Fail-closed: [`verify_release`] returns `false` on ANY mismatch — a tampered
-//! image, a wrong key, or a malformed/altered signature. The actuator releases
-//! ONLY on `true`.
+//! **No new crypto primitives** (HVCHAN §3): the SAME `ed25519-dalek` + `sha2`
+//! the verifier's attestation / audit-chain paths already use, applied at this
+//! seam. `#![forbid(unsafe_code)]`.
+//!
+//! **One canonical token format.** Both the digest and the token signature use
+//! **domain-separated, length-prefixed** payloads (the `audit_chain` /
+//! `compute_causal_record_hash` house style), so a governor-release digest or
+//! signature can never collide with — or be replayed as — an audit-chain or causal
+//! hash/signature. This crate is the SINGLE source of the release-token format:
+//! the heavy `kirra-verifier` root crate re-exports it from
+//! `src/governor_release.rs` (which adds only the root-local `governor_key_id`
+//! forensic helper), so there is exactly one digest/sign/verify implementation.
 
 #![forbid(unsafe_code)]
 
@@ -27,138 +40,244 @@ use sha2::{Digest, Sha256};
 
 use kirra_contract_channel::GovernorContractView;
 
-/// A governor release token: the Ed25519 signature (64 bytes) over the SHA-256
-/// digest of the validated contract image. The bytes an actuator checks before
-/// releasing a command.
+/// Domain tag for the contract digest (step 5). Distinct from every audit-chain
+/// / causal tag so the two hash spaces never collide.
+const DIGEST_DOMAIN: &[u8] = b"KIRRA-GOVERNOR-CONTRACT-DIGEST-V1";
+
+/// Domain tag for the release-token signing payload (step 6). Distinct again, so
+/// a release signature can never be reused as an audit signature and vice versa.
+const RELEASE_DOMAIN: &[u8] = b"KIRRA-GOVERNOR-RELEASE-V1";
+
+/// The release token: a digest of the approved command plus the governor's
+/// Ed25519 signature over it. Minimal and fixed-size (96 bytes on the wire); the
+/// actuator is supplied the governor verifying key out-of-band.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct ReleaseToken {
-    /// The raw Ed25519 signature over the image digest.
+    /// SHA-256 digest over the validated snapshot's canonical image (step 5).
+    pub digest: [u8; 32],
+    /// Ed25519 signature over the domain-separated digest payload (step 6).
     pub signature: [u8; 64],
 }
 
-/// SHA-256 over the canonical contract image (HVCHAN §3 step 5). Deterministic;
-/// the actuator recomputes it from the image it holds, so a tampered image
-/// yields a different digest and fails verification.
+/// Why the actuator refused to release (step 7). Both are fail-closed.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum ReleaseDenied {
+    /// The token's digest does not match the command about to be actuated — the
+    /// approval was for *different* bytes. (Substitution / stale token.)
+    DigestMismatch,
+    /// The signature does not verify against the governor key — forged or
+    /// tampered token, or wrong signer.
+    SignatureInvalid,
+}
+
+/// The digest payload (step 5): domain tag, then the length-prefixed canonical
+/// image. The image is fixed-size, but the length prefix keeps the encoding
+/// injective and consistent with the audit-chain discipline.
+fn digest_payload(view: &GovernorContractView) -> ([u8; 32 + 1024], usize) {
+    // Stack buffer sized to comfortably hold the domain tag + 8-byte length +
+    // the canonical image (176 bytes today). Returned with its used length.
+    let image = view.canonical_image();
+    let mut buf = [0u8; 32 + 1024];
+    let mut n = 0;
+    buf[n..n + DIGEST_DOMAIN.len()].copy_from_slice(DIGEST_DOMAIN);
+    n += DIGEST_DOMAIN.len();
+    buf[n..n + 8].copy_from_slice(&(image.len() as u64).to_le_bytes());
+    n += 8;
+    buf[n..n + image.len()].copy_from_slice(&image);
+    n += image.len();
+    (buf, n)
+}
+
+/// Step 5: the digest over the exact validated snapshot bytes. Deterministic;
+/// independent of the live shared region.
 #[must_use]
-pub fn digest_image(canonical_image: &[u8]) -> [u8; 32] {
+pub fn contract_digest(view: &GovernorContractView) -> [u8; 32] {
+    let (buf, n) = digest_payload(view);
     let mut hasher = Sha256::new();
-    hasher.update(canonical_image);
+    hasher.update(&buf[..n]);
     hasher.finalize().into()
 }
 
-/// Sign the digest of `canonical_image` with the governor's key (HVCHAN §3
-/// steps 5-6). The image is [`GovernorContractView::canonical_image`] of the
-/// exact validated snapshot; see [`sign_view`] for the typed convenience.
-///
-/// The image is a **fixed-length** `&[u8; CANONICAL_IMAGE_LEN]` so signing a
-/// wrong-length / non-canonical image is a COMPILE error, not a runtime footgun —
-/// reinforcing the "sign exactly the validated bytes" invariant at the type level.
-#[must_use]
-pub fn sign_release(
-    canonical_image: &[u8; kirra_contract_channel::CANONICAL_IMAGE_LEN],
-    signing_key: &SigningKey,
-) -> ReleaseToken {
-    let digest = digest_image(canonical_image);
-    ReleaseToken { signature: signing_key.sign(&digest).to_bytes() }
+/// The signing payload over a digest (step 6): domain tag, then the 32-byte
+/// digest length-prefixed. Domain separation prevents cross-protocol reuse.
+fn release_signing_payload(digest: &[u8; 32]) -> [u8; RELEASE_DOMAIN.len() + 8 + 32] {
+    let mut out = [0u8; RELEASE_DOMAIN.len() + 8 + 32];
+    let mut n = 0;
+    out[n..n + RELEASE_DOMAIN.len()].copy_from_slice(RELEASE_DOMAIN);
+    n += RELEASE_DOMAIN.len();
+    out[n..n + 8].copy_from_slice(&(32u64).to_le_bytes());
+    n += 8;
+    out[n..n + 32].copy_from_slice(digest);
+    out
 }
 
-/// Verify a release token against `canonical_image` and the governor's public
-/// key — the actuator's pre-release gate. **Fail-closed:** `false` on any
-/// mismatch (tampered image → different digest; wrong key; bad signature).
-///
-/// Fixed-length image (`&[u8; CANONICAL_IMAGE_LEN]`) for the same reason as
-/// [`sign_release`]: the actuator cannot accidentally verify over a non-canonical
-/// slice.
+/// Step 6: issue a release token binding `view` to the governor's approval, by
+/// signing its digest with `signing_key`. The governor calls this on an
+/// actuatable verdict, over the exact snapshot it validated (e.g.
+/// `kirra_core::contract_consumer::GovernorCycle::view_to_sign`).
 #[must_use]
+pub fn issue_release_token(view: &GovernorContractView, signing_key: &SigningKey) -> ReleaseToken {
+    let digest = contract_digest(view);
+    let signature = signing_key.sign(&release_signing_payload(&digest));
+    ReleaseToken { digest, signature: signature.to_bytes() }
+}
+
+/// Step 7: the actuator's verify-before-release gate. `actuating_view` is the
+/// command the actuator is about to drive; `governor_vk` is the trusted governor
+/// key. Returns `Ok(())` only if the token's digest matches the actuating command
+/// **and** the signature verifies — otherwise a fail-closed [`ReleaseDenied`].
 pub fn verify_release(
-    canonical_image: &[u8; kirra_contract_channel::CANONICAL_IMAGE_LEN],
     token: &ReleaseToken,
-    verifying_key: &VerifyingKey,
-) -> bool {
-    let digest = digest_image(canonical_image);
-    verifying_key
-        .verify_strict(&digest, &Signature::from_bytes(&token.signature))
-        .is_ok()
+    actuating_view: &GovernorContractView,
+    governor_vk: &VerifyingKey,
+) -> Result<(), ReleaseDenied> {
+    // (a) The token must approve exactly the bytes about to be actuated.
+    let expected = contract_digest(actuating_view);
+    if token.digest != expected {
+        return Err(ReleaseDenied::DigestMismatch);
+    }
+    // (b) The signature must verify against the governor key, over the same
+    // digest. `verify_strict` is the attestation path (rejects malleable / small-
+    // order points). A tampered digest field fails here even if (a) somehow passed.
+    let sig = Signature::from_bytes(&token.signature);
+    if governor_vk
+        .verify_strict(&release_signing_payload(&token.digest), &sig)
+        .is_err()
+    {
+        return Err(ReleaseDenied::SignatureInvalid);
+    }
+    Ok(())
 }
 
-/// Sign the validated [`GovernorContractView`] (its canonical image). The
-/// governor calls this on an actuatable verdict, over the exact snapshot it
-/// validated (e.g. `kirra_core::contract_consumer::GovernorCycle::view`).
-#[must_use]
-pub fn sign_view(view: &GovernorContractView, signing_key: &SigningKey) -> ReleaseToken {
-    sign_release(&view.canonical_image(), signing_key)
-}
+impl ReleaseToken {
+    /// Canonical 96-byte wire form: `digest(32) || signature(64)`.
+    #[must_use]
+    pub fn to_bytes(&self) -> [u8; 96] {
+        let mut out = [0u8; 96];
+        out[..32].copy_from_slice(&self.digest);
+        out[32..].copy_from_slice(&self.signature);
+        out
+    }
 
-/// Verify a release token against a [`GovernorContractView`] — the actuator's
-/// gate over the typed view. Fail-closed, as [`verify_release`].
-#[must_use]
-pub fn verify_view(
-    view: &GovernorContractView,
-    token: &ReleaseToken,
-    verifying_key: &VerifyingKey,
-) -> bool {
-    verify_release(&view.canonical_image(), token, verifying_key)
+    /// Parse the canonical wire form. (No verification — that is [`verify_release`].)
+    #[must_use]
+    pub fn from_bytes(bytes: &[u8; 96]) -> Self {
+        let mut digest = [0u8; 32];
+        let mut signature = [0u8; 64];
+        digest.copy_from_slice(&bytes[..32]);
+        signature.copy_from_slice(&bytes[32..]);
+        Self { digest, signature }
+    }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
 
-    fn keypair(seed: u8) -> (SigningKey, VerifyingKey) {
-        // Deterministic (no RNG) — same pattern as the attestation tests.
-        let sk = SigningKey::from_bytes(&[seed; 32]);
-        let vk = sk.verifying_key();
-        (sk, vk)
+    fn governor_key() -> SigningKey {
+        SigningKey::from_bytes(&[42u8; 32])
     }
 
-    fn image() -> [u8; kirra_contract_channel::CANONICAL_IMAGE_LEN] {
-        GovernorContractView::new_command(2, 1, 0, u64::MAX / 2, b"steer:1.0")
-            .unwrap()
-            .canonical_image()
+    fn view(seq: u64, cmd: &[u8]) -> GovernorContractView {
+        GovernorContractView::new_command(2, seq, 100, 10_000, cmd).unwrap()
     }
 
     #[test]
-    fn sign_then_verify_accepts() {
-        let (sk, vk) = keypair(1);
-        let img = image();
-        let token = sign_release(&img, &sk);
-        assert!(verify_release(&img, &token, &vk));
+    fn honest_token_releases() {
+        let sk = governor_key();
+        let v = view(1, b"steer:1.5");
+        let token = issue_release_token(&v, &sk);
+        assert_eq!(verify_release(&token, &v, &sk.verifying_key()), Ok(()));
     }
 
     #[test]
-    fn a_tampered_image_is_rejected() {
-        let (sk, vk) = keypair(1);
-        let mut img = image();
-        let token = sign_release(&img, &sk);
-        img[48] ^= 0xFF; // flip a command byte AFTER signing
-        assert!(!verify_release(&img, &token, &vk));
+    fn digest_binds_the_exact_command_bytes() {
+        // A token issued for one command must NOT release a different command,
+        // even a one-byte change — the digest mismatch is caught.
+        let sk = governor_key();
+        let approved = view(1, b"steer:1.5");
+        let token = issue_release_token(&approved, &sk);
+
+        let tampered = view(1, b"steer:9.9");
+        assert_eq!(
+            verify_release(&token, &tampered, &sk.verifying_key()),
+            Err(ReleaseDenied::DigestMismatch)
+        );
     }
 
     #[test]
-    fn a_wrong_public_key_is_rejected() {
-        let (sk, _) = keypair(1);
-        let (_, other_vk) = keypair(2);
-        let img = image();
-        let token = sign_release(&img, &sk);
-        assert!(!verify_release(&img, &token, &other_vk));
+    fn any_header_field_change_changes_the_digest() {
+        // The digest covers the whole canonical image, not just the payload: a
+        // changed sequence (replay number) or deadline yields a different digest.
+        let sk = governor_key();
+        let token = issue_release_token(&view(1, b"go"), &sk);
+        assert_eq!(
+            verify_release(&token, &view(2, b"go"), &sk.verifying_key()),
+            Err(ReleaseDenied::DigestMismatch),
+            "a different sequence is a different approved command"
+        );
     }
 
     #[test]
-    fn a_tampered_token_is_rejected() {
-        let (sk, vk) = keypair(1);
-        let img = image();
-        let mut token = sign_release(&img, &sk);
-        token.signature[0] ^= 0xFF;
-        assert!(!verify_release(&img, &token, &vk));
+    fn wrong_governor_key_is_rejected() {
+        let real = governor_key();
+        let imposter = SigningKey::from_bytes(&[7u8; 32]);
+        let v = view(1, b"go");
+        // Token signed by the imposter; actuator trusts the real governor key.
+        let token = issue_release_token(&v, &imposter);
+        assert_eq!(
+            verify_release(&token, &v, &real.verifying_key()),
+            Err(ReleaseDenied::SignatureInvalid)
+        );
     }
 
     #[test]
-    fn sign_view_agrees_with_sign_release_over_the_canonical_image() {
-        let (sk, vk) = keypair(3);
-        let view = GovernorContractView::new_command(4, 2, 0, u64::MAX / 2, b"steer:2.0").unwrap();
-        let token = sign_view(&view, &sk);
-        assert!(verify_view(&view, &token, &vk));
-        // Same as signing the raw canonical image.
-        assert_eq!(token, sign_release(&view.canonical_image(), &sk));
+    fn tampered_signature_is_rejected() {
+        let sk = governor_key();
+        let v = view(1, b"go");
+        let mut token = issue_release_token(&v, &sk);
+        token.signature[0] ^= 0x01;
+        assert_eq!(
+            verify_release(&token, &v, &sk.verifying_key()),
+            Err(ReleaseDenied::SignatureInvalid)
+        );
+    }
+
+    #[test]
+    fn tampered_digest_field_is_rejected() {
+        // Flipping the token's digest to match a substituted command does not
+        // help: the signature was over the original digest, so verify fails.
+        let sk = governor_key();
+        let v = view(1, b"go");
+        let mut token = issue_release_token(&v, &sk);
+        let substitute = view(1, b"no");
+        token.digest = contract_digest(&substitute); // now matches `substitute`
+        assert_eq!(
+            verify_release(&token, &substitute, &sk.verifying_key()),
+            Err(ReleaseDenied::SignatureInvalid),
+            "digest matches the substitute, but the signature does not cover it"
+        );
+    }
+
+    #[test]
+    fn wire_roundtrip_preserves_the_token() {
+        let sk = governor_key();
+        let token = issue_release_token(&view(3, b"abc"), &sk);
+        assert_eq!(ReleaseToken::from_bytes(&token.to_bytes()), token);
+    }
+
+    #[test]
+    fn release_payload_is_domain_separated_from_a_bare_digest() {
+        // The signed payload is NOT the bare digest; a signature is bound to the
+        // release domain and cannot be reused where a bare-digest signature is
+        // expected. (Guards against cross-protocol signature reuse.)
+        let sk = governor_key();
+        let v = view(1, b"go");
+        let token = issue_release_token(&v, &sk);
+        let bare = Signature::from_bytes(&token.signature);
+        assert!(
+            sk.verifying_key().verify_strict(&token.digest, &bare).is_err(),
+            "the signature must not verify over the bare digest (no domain tag)"
+        );
     }
 }
